requirements.txt
blinker==1.6.2
boto3==1.28.34
botocore==1.31.34
click==8.1.7
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.16.2
python-dotenv==1.0.0
